Create test project Simplify the way to write text to the HTTP Response
