+++ nac.WebServer/lib/HttpHelper.cs
@@ -8,22 +8,7 @@ namespace nac.WebServer.lib;
 
 public static class HttpHelper
 {
-    public static void WriteTextResponse(HttpListenerResponse response, string contentType, string text)
-    {
-        var buf = Encoding.UTF8.GetBytes(text);
-        response.ContentType = contentType;
-        response.ContentLength64 = buf.Length;
-        response.OutputStream.Write(buf, 0, buf.Length);
-    }
-
-    public static void WriteBinaryResponse(HttpListenerResponse response, string contentType, byte[] data)
-    {
-        response.ContentType = contentType;
-        response.ContentLength64 = data.Length;
-        response.OutputStream.Write(data, 0, data.Length);
-    }
-
-
+    
     public static models.FileServeData getFileServeInfo(string extension)
     {
         var info = new models.FileServeData();
